Modified USER and COURSE entities and added controllers for COURSE
Added visibility specifiers and 'remove' on cascade

+++ server/src/controllers/courses.ts
@@ -8,45 +8,67 @@ import {compareClasses, alterKeys, processCollisions, listCollisions} from "../s
 const logger = require('../utils/logger')
 const coursesRouter = require('express').Router()
 
-coursesRouter.delete('/deleteDummy', async (request: Request, response: Response) => {
+coursesRouter.post('/', async (request: Request, response: Response) => {
+  const body = request.body
+
+  const newCourse: Course = {
+    courseKey: body.courseKey,
+    name: body.name,
+    classes: []
+  }
+
+  await getConnection()
+  .createQueryBuilder()
+  .insert()
+  .into(Course)
+  .values(newCourse)
+  .execute()
+  .then(() => {
+    return response.status(200).json({
+      status: "success"
+    })
+  })
+  .catch(error => {
+    logger.error(error)
+    return response.status(500).json({
+      status: "failure",
+      message: error.message
+    })
+  });
+})
+
+coursesRouter.delete('/:courseKey', async (request: Request, response: Response) => {
+  const courseKey = request.params.courseKey
+
+  const connection = await getConnection();
+  const courseRepository = connection.getRepository(Course)
+  const course = await courseRepository.findOne({courseKey});
+
+  if (course == undefined) {
+    return response.status(500).json({
+      status: "failure",
+      message: "course not found"
+    })
+  }
+
   await getConnection()
     .createQueryBuilder()
     .delete()
     .from(Course)
-    .where("course.id > :id", { id: 0 })
+    .where("courseKey = :courseKey", {courseKey})
     .execute()
     .then(() => {
-      return response.json({status: "jest w pyte"})
+      return response.status(200).json({
+        status: "success"
+      })
     })
-    .catch(error => logger.error(error));
-})
-
-coursesRouter.post('/addDummy', async (request: Request, response: Response) => {
-    const connection = await getConnection();
-    const userRepository = connection.getRepository(User)
-
-    const usr = await userRepository.findOne({username: "2"});
-    const css = [
-        {
-            name: "Zastosowania informatyki w gospodarce",
-            user: usr
-        },
-        {
-            name: "Matematyka",
-            user: usr
-        }
-    ]
-
-    await getConnection()
-    .createQueryBuilder()
-    .insert()
-    .into(Course)
-    .values(css)
-    .execute()
-    .then(() => {
-      return response.json(css)
-    })
-    .catch(error => logger.error(error));
+    .catch(error => {
+      logger.error(error)
+      return response.status(500).json({
+        status: "failure",
+        message: error.message
+      })
+    });
 })
 
 // PZD-10
@@ -59,39 +81,133 @@ coursesRouter.get('/names', async (request: Request, response: Response) => {
     .from(Course, "course")
     .execute()
     .then(items => {
-      return response.status(200).json(alterKeys(items, "course").map((i: { name: any; }) => i.name))
+      return response
+      .status(200)
+      .json(alterKeys(items, "course").map((item: { name: string; }) => item.name))
     })
     .catch(error => {
       logger.error(error)
       return response.status(500).json({
-          success: false,
-          status: "nk rzutnik"
+          status: "failure",
+          message: error.message
       })
     });
 })
 
-// courses of user id=2
-coursesRouter.get('/2', async (request: Request, response: Response) => {
-    const connection = await getConnection();
-    const userRepository = connection.getRepository(User)
-    const coursesRepository = connection.getRepository(Course)
+// coursesOfUser(username) -> Courses
+coursesRouter.get('/user/:username', async (request: Request, response: Response) => {
+  const username = request.params.username
 
-    const user = await userRepository.findOne({username: "2"})
-    const courses = await coursesRepository.find({user});
+  const connection = await getConnection();
+  const userRepository = connection.getRepository(User)
+  const user = await userRepository.findOne({username});
 
-    return response.json(courses)
+  if (user == undefined) {
+    return response.status(500).json({
+      status: "failure",
+      message: "specified user does not exist"
+    })
+  }
+
+  await getConnection()
+    .createQueryBuilder()
+    .select("course")
+    .from(Course, "course")
+    .where("userUsername = :user", {user: user.username})
+    .execute()
+    .then(items => {
+      return response
+      .status(200)
+      .json(alterKeys(items, "course"))
+    })
+    .catch(error => {
+      logger.error(error)
+      return response.status(500).json({
+          status: "failure",
+          message: error.message
+      })
+    });
 })
 
-//all courses
+coursesRouter.get('/:courseKey', async (request: Request, response: Response) => {
+  const courseKey = request.params.courseKey
+
+  await getConnection()
+    .createQueryBuilder()
+    .select("course")
+    .from(Course, "course")
+    .where("courseKey = :courseKey", {courseKey})
+    .execute()
+    .then(items => {
+      return response
+      .status(200)
+      .json(alterKeys(items, "course"))
+    })
+    .catch(error => {
+      logger.error(error)
+      return response.status(500).json({
+          status: "failure",
+          message: error.message
+      })
+    });
+})
+
 coursesRouter.get('/', async (request: Request, response: Response) => {
     const connection = await getConnection();
     const coursesRepository = connection.getRepository(Course)
 
     const courses = await coursesRepository.find();
 
-    return response.json(courses)
+    return response.status(200).json(courses)
 })
 
+coursesRouter.put('/:courseKey', async (request: Request, response: Response) => {
+  const body = request.body
+  const courseKey = request.params.courseKey
+
+  const connection = await getConnection();
+  const courseRepository = connection.getRepository(Course)
+  const course = await courseRepository.findOne({courseKey});
+
+  const userRepository = connection.getRepository(User)
+  const user = await userRepository.findOne({username: body.user.username});
+
+  if (course == undefined) {
+    return response.status(500).json({
+      status: "failure",
+      message: "course not found"
+    })
+  }
+  if (user == undefined) {
+    return response.status(500).json({
+      status: "failure",
+      message: "specified user does not exist"
+    })
+  }
+
+  await getConnection()
+    .createQueryBuilder()
+    .update(Course)
+    .set({
+        name: body.name,
+        courseKey: body.courseKey,
+        user: user
+    })
+    .where("courseKey = :courseKey", {courseKey})
+    .execute()
+    .then(() => {
+      return response.status(200).json({
+        status: "success"
+      })
+    })
+    .catch(error => {
+      logger.error(error)
+      return response.status(500).json({
+        status: "failure",
+        message: error.message
+      })
+    });
+})
